feat: 메뉴(음료) 관련 타입 menu 폴더로 이동
