+++ src/components/todayCaffeine/HistoryList.tsx
@@ -1,4 +1,4 @@
-import { Menu } from '@/types/home/menu';
+import { Menu } from '@/types/menu/menu';
 import DrinkListItem from '../common/drink/DrinkListItem';
 
 const HistoryList = ({ drinkList }: { drinkList: Menu[] }) => {
