+++ src/constants/menu/menuList.ts
@@ -1,0 +1,1 @@
+export const MENU_PER_PAGE = 12;
