+++ src/types/home/user.ts
@@ -1,4 +1,4 @@
-import { Menu } from './menu';
+import { Menu } from '../menu/menu';
 
 export interface UserCaffeineData {
   nickname: string;
