+++ src/types/menu/menu.ts
@@ -1,3 +1,5 @@
+import { Nutrient, NutrientLevel } from './nutrient';
+
 export interface Menu {
   menuNo: number;
   menuName: string;
@@ -26,24 +28,3 @@ export interface ComparedMenu extends Menu {
   kcal: string;
   volume: string;
 }
-
-export interface Nutrient {
-  kcal: string;
-  sugar: string;
-  salt: string;
-  protein: string;
-  satFat: string;
-}
-
-export type NutrientLevelType = '높음' | '보통' | '낮음';
-export interface NutrientLevel {
-  kcalLevel: NutrientLevelType;
-  sugarLevel: NutrientLevelType;
-  saltLevel: NutrientLevelType;
-  proteinLevel: NutrientLevelType;
-  satFatLevel: NutrientLevelType;
-}
-
-export type Entries<T> = {
-  [K in keyof T]: [K, T[K]];
-}[keyof T][];
